DoctorWho/sepsis_challenge_s6/requirements.txt
absl-py==0.7.1
astor==0.8.0
bleach==1.5.0
colorama==0.4.1
decorator==4.4.0
docopt==0.6.2
gast==0.2.2
grpcio==1.22.0
html5lib==0.9999999
joblib==0.13.2
jsonpickle==0.9.6
Markdown==3.1.1
munch==2.3.2
networkx==2.3
numpy==1.16.4
packaging==19.0
pandas==0.24.2
protobuf==3.8.0
py-cpuinfo==5.0.0
pymongo==3.8.0
pyparsing==2.4.0
python-dateutil==2.8.0
pytz==2019.1
sacred==0.7.5
scikit-learn==0.21.2
scipy==1.3.0
six==1.12.0
sklearn==0.0
tensorflow-gpu==1.10.0
termcolor==1.1.0
toposort==1.5
Werkzeug==0.15.4
wrapt==1.11.2
